+++ xagg/__init__.py
@@ -3,5 +3,5 @@
 # happening as methods to the classes that are exported from those
 # two functions)
 from .wrappers import pixel_overlaps
-from .aux import (normalize,fix_ds,get_bnds,subset_find)
+from .auxfuncs import (normalize,fix_ds,get_bnds,subset_find)
 from .core import (aggregate,read_wm)
