+++ xagg/auxfuncs.py
